builtin/repack: fix geometric repacks with promisor remotes
When repacking a repository with promisor remotes git-repack(1) knows to
pass "--exclude-promisor-objects" to git-pack-objects(1). This option
ensures that the new pack will not contain any promised object that do
not yet exist locally.

This command line option is incompatible with "--stdin-packs": the
latter option enables the rev-walk-based machinery to figure out which
objects to add to the pack, whereas the former tells git-pack-objects(1)
to merge all packs passed via stdin into one large pack. As we do not
know to filter those packs via the passed-in revisions it is clear that
at the current point in time nothing sensible comes out of combining
these two options.

But there is one case where git-repack(1) decides to pass both options:
when performing a geometric repack we always pass "--stdin-packs" to
identify the packs that should be merged. So if one performs a geometric
repack in a partial clone we'll end up with both options, and that
causes the repack to fail.

Fix this issue by never passing "--exclude-promisor-objects" when we
have a geometric split factor. We don't need the option anyway when
doing a geometric repack as we will only ever pack loose objects or
merge multiple packs. And neither of those cases can yield a promisor
object.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>